Removed pre-1.0 deprecations warnings
